int/float

+++ payCalculator.py
@@ -2,7 +2,7 @@ def calculatePay():
     # Implement your solution in between the two comment blocks
     print("calculating pay")
     # This first line is provided for you
-    hrs = float(input ("Enter Hours: "))
+    hrs = int(input ("Enter Hours: "))
     rate = float (input("Enter Rate: "))
     pay = hrs * rate
     print (pay)
